+++ bench/rsc/jvm/src/main/scala/rsc/bench/RscScope.scala
@@ -1,0 +1,60 @@
+// Copyright (c) 2017 Twitter, Inc.
+// Licensed under the Apache License, Version 2.0 (see LICENSE.md).
+package rsc.bench
+
+import java.util.concurrent.TimeUnit
+import org.openjdk.jmh.annotations._
+import org.openjdk.jmh.annotations.Mode._
+import rsc.bench.RscScope._
+import rsc.semantics._
+
+object RscScope {
+  @State(Scope.Benchmark)
+  class BenchmarkState extends RscFixtures with FileFixtures
+}
+
+trait RscScope {
+  def runImpl(bs: BenchmarkState): Unit = {
+    val compiler = bs.mkCompiler("-Ystop-after:scope", bs.re2sRscFiles)
+    compiler.run()
+    val problems = compiler.reporter.problems
+    if (problems.nonEmpty) {
+      problems.foreach(println)
+      sys.error("scope failed")
+    }
+  }
+}
+
+@BenchmarkMode(Array(SingleShotTime))
+@OutputTimeUnit(TimeUnit.MILLISECONDS)
+@Fork(value = 128, jvmArgs = Array("-Xms2G", "-Xmx2G"))
+class ColdRscScope extends RscScope {
+  @Benchmark
+  def run(bs: BenchmarkState): Unit = {
+    runImpl(bs)
+  }
+}
+
+@BenchmarkMode(Array(SampleTime))
+@OutputTimeUnit(TimeUnit.MILLISECONDS)
+@Warmup(iterations = 10, time = 1, timeUnit = TimeUnit.SECONDS)
+@Measurement(iterations = 10, time = 1, timeUnit = TimeUnit.SECONDS)
+@Fork(value = 1, jvmArgs = Array("-Xms2G", "-Xmx2G"))
+class WarmRscScope extends RscScope {
+  @Benchmark
+  def run(bs: BenchmarkState): Unit = {
+    runImpl(bs)
+  }
+}
+
+@BenchmarkMode(Array(SampleTime))
+@OutputTimeUnit(TimeUnit.MILLISECONDS)
+@Warmup(iterations = 10, time = 10, timeUnit = TimeUnit.SECONDS)
+@Measurement(iterations = 10, time = 10, timeUnit = TimeUnit.SECONDS)
+@Fork(value = 3, jvmArgs = Array("-Xms2G", "-Xmx2G"))
+class HotRscScope extends RscScope {
+  @Benchmark
+  def run(bs: BenchmarkState): Unit = {
+    runImpl(bs)
+  }
+}
